The mail compiler is less opinionated.
A new method `make_bootstrap_mail` is introduced aimed to replace the
method `bootstrap_mail` that has been kept for compatibility.

The new method does not impose the layout name, furthermore it allows the
developer to write a textual version of the mail without beign compelled
to use the auto generated one make by the `premailer-rails` gem.