Fix bug where SASS load paths became nested arrays
By adding `BootstrapEmail.config.sass_load_paths`, which is an array, to `SassC.load_paths`, the two arrays are not joined but the first is added to the latter as a new array element. This leads to a nested array.

Nested array load paths can lead to exceptions during sass compilation, for instance when using Dart Sass via the Gem `sass-embedded`.