force tint and shade to return colors that will be 6-digit hex codes
This addresses https://github.com/bootstrap-email/bootstrap-email/issues/280

Color values should now be like `bgcolor="CFE2FF"` instead of `bgcolor="rgb(206.6, 226, 254.6)"` for `bg-blue-100`. This isn't perfect as the value is missing the `#` - the browser will accept this color value, but I'm not sure about all email clients.

I could not figure out a way to get the sass compiler to return a color value with a # or otherwise avoid an rgb function. `ie-hex-str` seems like it could be useful, but that returns a string (with an alpha value) and the sass compiler expects a color object, not a string.

An alternate approach would be to replace all the function-defined colors in `_colors.scss` with their explicit values. E.g. `"blue-100": tint-color($blue, 80%)` to `"blue-100": #CFE2FF` - this works but doesn't have the same elegance of the existing approach.